sentence-gen-api/requirements.txt
absl-py==2.0.0
accelerate==0.28.0
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
amqp==5.2.0
annotated-types==0.5.0
ansi2html==1.9.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.6.2
blis==0.7.10
bs4==0.0.2
build==1.1.1
cachelib==0.10.2
cachetools==5.3.1
catalogue==2.0.9
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
colorlover==0.3.0
confection==0.1.1
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.2
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.7
dash==2.13.0
dash-ag-grid==2.4.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
Django==4.2.7
dnspython==2.4.2
docker==7.0.0
docxcompose==1.4.0
docxtpl==0.16.7
ecdsa==0.18.0
effdet==0.4.1
elastic-transport==8.4.0
elasticsearch==8.9.0
email-validator==2.1.0.post1
emoji==2.10.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
extra-streamlit-components==0.1.60
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.13.1
filetype==1.2.0
FLAML==2.1.2
Flask==2.2.0
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Session==0.5.0
flatbuffers==23.5.26
fonttools==4.49.0
frozenlist==1.3.3
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==2.0.2
grpcio==1.62.1
grpcio-tools==1.62.1
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
icecream==2.1.3
idna==3.6
importlib-metadata==6.6.0
importlib-resources==6.1.1
incremental==22.10.0
iopath==0.1.10
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jose==1.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-specifications==2023.11.2
jwt==1.3.1
keras==2.14.0
keras-models==0.0.7
kiwisolver==1.4.5
kombu==5.3.5
kubernetes==29.0.0
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-experimental==0.0.54
langchain-openai==0.0.6
langchain-text-splitters==0.0.1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.1.26
layoutparser==0.3.4
libclang==16.0.6
lxml==5.1.0
Markdown==3.5.1
markdown-it-py==2.2.0
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-enum==1.5.1
matplotlib==3.7.2
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.26.4
oauthlib==3.2.2
ollama==0.1.7
omegaconf==2.3.0
onnx==1.15.0
# onnxruntime==1.15.1
openai==1.12.0
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parsel==1.8.1
passlib==1.7.4
pathlib==1.0.1
pathy==0.10.2
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.11.0
pillow==10.2.0
pillow_heif==0.15.0
plotly==5.18.0
portalocker==2.8.2
posthog==3.5.0
preshed==3.0.8
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.23.4
psutil==5.9.8
psycopg2==2.9.9
pulsar-client==3.4.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyautogen==0.2.19
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydeck==0.8.1b0
PyDispatcher==2.0.7
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pymssql==2.2.11
PyMySQL==1.0.3
pyodbc==4.0.34
pyOpenSSL==24.0.0
pyparsing==3.0.9
pypdf==4.0.1
PyPDF2==3.0.1
pypdfium2==4.27.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2024.2.7
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
qdrant-client==1.8.0
queuelib==1.6.2
rapidfuzz==3.6.1
rarfile==4.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.3.5
rpds-py==0.13.2
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.2
scipy==1.10.1
Scrapy==2.11.1
sentence-transformers==2.5.1
seqeval==1.2.2
service-identity==24.1.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.20
sqlparse==0.4.4
srsly==2.4.7
starlette==0.27.0
streamlit==1.29.0
streamlit-authenticator==0.2.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.1.12
threadpoolctl==3.2.0
tiktoken==0.6.0
timm==0.9.12
tldextract==5.1.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
tornado==6.3.2
tqdm==4.66.2
transformers==4.37.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.0.1
ujson==5.8.0
unstructured==0.12.6
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.24.0.post1
validators==0.20.0
vine==5.1.0
w3lib==2.1.2
wasabi==0.10.1
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.2.3
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.1.9
yarl==1.9.2
zipp==3.17.0
zope.interface==6.2
